Unify project_type checks and templatetags
