+++ meinberlin/apps/embed/templates/meinberlin_embed/embed.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-{% load static i18n extproject_tags %}
+{% load static i18n meinberlin_project_tags %}
 {% get_current_language as LANGUAGE_CODE %}
 <html class="embed" lang="{{ LANGUAGE_CODE }}">
 <head>
