+++ meinberlin/apps/projects/templates/meinberlin_projects/includes/project_list_tile.html
@@ -1,20 +1,20 @@
-{% load i18n project_tags extproject_tags contrib_tags thumbnail static %}
+{% load i18n project_tags meinberlin_project_tags contrib_tags thumbnail static %}
 {% project_tile_image project as project_image %}
 {% project_tile_image_copyright project as project_image_copyright %}
 
-<li class="tile {% if project.projectcontainer %}tile--container{% endif %}">
+<li class="tile {% if project|is_container %}tile--container{% endif %}">
     <div class="tile__head">
         <a tabindex="-1" class="tile__image tile__image--shadowed"
            aria-labelledby="project-title-{{ project.pk }}"
            {% if project_image %}
                style="background-image: url({{ project_image|thumbnail_url:'project_thumbnail' }})"
            {% endif %}
-           href="{% extproject_url project 'project-detail' %}" target="{{ project|is_external|yesno:"_blank,_self" }}">
+           href="{{ project|project_url }}" target="{{ project|is_external|yesno:"_blank,_self" }}">
         </a>
         {% if project_image_copyright %}
             <div class="tile__image__copyright copyright">© {{ project_image_copyright }}</div>
         {% endif %}
-        {% if not project.projectcontainer %}
+        {% if not project|is_container %}
             {% if project|is_external %}
                 <img src="{% static 'images/external_project.svg' %}" alt="{% trans 'External project' %}" class="tile__badge">
             {% else %}
@@ -38,12 +38,12 @@
 
     <div class="tile__body">
         <h3 class="tile__title" id="project-title-{{ project.pk }}">
-            <a href="{% extproject_url project 'project-detail' %}" target="{{ project|is_external|yesno:"_blank,_self" }}">{{ project.name }}</a>
+            <a href="{{ project|project_url }}" target="{{ project|is_external|yesno:"_blank,_self" }}">{{ project.name }}</a>
         </h3>
 
         <p class="tile__text">{{ project.description }}</p>
 
-        {% if project.projectcontainer %}
+        {% if project|is_container %}
             <p class="tile__hint">
                 {% blocktrans with count=project.projectcontainer.projects.count active_count=project.projectcontainer.active_projects.count %}Currently {{ active_count }} of {{ count }} projects are active.{% endblocktrans %}
             </p>
